+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog of Commits
 Made to document team contributions (who worked on what).
 
+
+## - Joses 14/05/2026 11:17:00
+### Added
+- Added InputTime function
+- Added MenuOption switch case
+### Updated
+- Updated Timer function (Input to InputTime)
+### Removed
+- Removed Hewwo Wowld Function
+
+
+## - Farrell 13/05/2026 12:37:05
+### Added
+- Added Menu, Timer, and ASCII art Function
+
 ## - Joses 09/05/2026 00:15:00
 ### Updated
 - Updated CHANGELOG.md